+++ src/main/kotlin/com/wiggglez/kt_stockdata_lib/historicchartdata/dataobjects/candle/Candle.kt
@@ -4,10 +4,10 @@ import java.util.*
 
 data class Candle(
     val datetime: Date,
-    val timestamp: Int,
+    val timestamp: Long,
     val open: Double,
     val high: Double,
     val low: Double,
     val close: Double,
-    val volume: Int
+    val volume: Long
 )
